[Merge record] part 1, refactor animation
- Add SortAnimationGemini.ts
- Add SortReplyParser.ts
- move the function about the animaiton sort logic
    in src/extension.ts into above 2 file
- let parseNodeSequence() return normal order;
    without really sorting
